Implement mongo database
