+++ package.json
@@ -4,8 +4,8 @@
   "description": "A url shortening app",
   "main": "express_server.js",
   "scripts": {
-    "start": "node express_server.js",
-    "dev": "nodemon express_server.js"
+    "start": "node server/express_server.js",
+    "dev": "nodemon server/express_server.js"
   },
   "author": "",
   "license": "ISC",
@@ -15,7 +15,8 @@
     "cookie-session": "^2.0.0-beta.3",
     "dotenv": "^6.2.0",
     "ejs": "^2.6.1",
-    "express": "^4.16.4"
+    "express": "^4.16.4",
+    "mongodb": "^3.1.13"
   },
   "engines": {
     "node": "9.11.2"
